add spinner

+++ src/Pages/MyReview/MyReview.js
@@ -1,6 +1,8 @@
 import React, { useContext, useEffect, useState } from "react";
 import { AuthProvider } from "../../Context/AuthContext";
 import ReviewCard from "./ReviewCard/ReviewCard";
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 const MyReview = () => {
   const [userReviews, setUserReviews] = useState([]);
@@ -14,6 +16,11 @@ const MyReview = () => {
   }, [user?.uid]);
 
   const handleDelete = (id) => {
+    const agree = window.confirm("Aye you sure to delete this review?");
+    if (!agree) {
+      return;
+    }
+
     fetch(`http://localhost:5000/reviews/${id}`, {
       method: "DELETE",
     })
@@ -21,8 +28,11 @@ const MyReview = () => {
       .then((data) => {
         console.log(data);
         if (data.deletedCount > 0) {
+          toast.success("Successfully Deleted");
           const remaining = userReviews.filter((review) => review._id !== id);
           setUserReviews(remaining);
+        } else {
+          return;
         }
       });
   };
@@ -32,14 +42,32 @@ const MyReview = () => {
         Your Reviews
       </h2>
       <div>
-        {userReviews.map((singleReview) => (
+        {userReviews.length === 0 ? (
+          <div className="w-3/5 mx-auto h-80 my-auto flex justify-center items-center">
+            <h3 className="text-3xl  text-center text-yellow-500 font-bold">
+              No reviews were added
+            </h3>
+          </div>
+        ) : (
+          <>
+            {userReviews.map((singleReview) => (
+              <ReviewCard
+                key={singleReview._id}
+                singleReview={singleReview}
+                handleDelete={handleDelete}
+              ></ReviewCard>
+            ))}
+          </>
+        )}
+        {/* {userReviews.map((singleReview) => (
           <ReviewCard
             key={singleReview._id}
             singleReview={singleReview}
             handleDelete={handleDelete}
           ></ReviewCard>
-        ))}
+        ))} */}
       </div>
+      <ToastContainer />
     </div>
   );
 };
